requirements.txt
fastapi==0.61.0
h5py==2.10.0
numpy==1.18.5
Pillow==7.2.0
requests==2.24.0
torch==1.6.0
torchvision==0.7.0